QOLOE-380 split navbar json for easy diffing
