+++ src/components/bs5/navbar/navbar.varient.coBrand.data.json
@@ -1,0 +1,27 @@
+{
+  "metadata": {
+    "target_url": "metadata-index.html",
+    "options": {
+      "showHomeTitleSmallDevices": true,
+      "icon-root": "/assets/img/svg-icons.svg",
+      "home-active": true
+    }
+  },
+  "navigation": [
+    {
+      "title": "What's really in vapes?",
+      "icon": "",
+      "target_url": "index.html"
+    },
+    {
+      "title": "I want to quit",
+      "icon": "",
+      "target_url": "index.html"
+    },
+    {
+      "title": "Information for parents and carers",
+      "icon": "",
+      "target_url": "index.html"
+    }
+  ]
+}
